Added red panda photo, fixed red panda photo, and changed background
Altered the css to change the background to red. Forgot to add a commt for the red panda before this commit- Added the red panda, the site will show a landscape photo, unless the viewport is 600px or less. My attempt to change the background didnt work, still working on that